Implement search field as setting with album invalidation
- Added search field to Settings type and persistence
- Updated RPC calls to use single string filter instead of string[]
- Modified media() and mediaCount() to split search string into terms
- Updated monitorAlbums() to accept searchFilter parameter
- Added search setting change listener to invalidate albums
- Updated getAlbumContents() to use search setting instead of text filter
- Regenerated RPC client with new signatures

Prompt: Make the search field a setting, that is then used in rpc media() calls. Invalidate the list of albums when the search field changes. If not empty, call monitorAlbums() again, passing the search field value as a filter. Change the filter argument in rpc calls into a single string.